ucarp: add initscripts-functions as runtime dependency when using systemd
Issue: LIN1019-2695

The ucarp.service invokes script /usr/libexec/ucarp to start/stop the
ucarp service. But the /etc/init.d/functions file which is required by
the script is not installed by default when using systemd. Explicitly
set the initscripts-functions package as the runtime dependency when
using systemd.

(LOCAL REV: NOT UPSTREAM) -- sent to yocto on 20190924

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>